requirements.txt
aniso8601==9.0.1
click==8.0.1
Flask==2.0.1
Flask-RESTful==0.3.9
future==0.18.2
iso8601==0.1.16
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pyserial==3.5
pytz==2021.1
PyYAML==5.4.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
serial==0.0.97
six==1.16.0
Werkzeug==2.0.1
